+++ NEWS.md
@@ -5,11 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.15.23] unreleased
+## [0.15.24] unreleased
 
 ### Added
 
 * A new tutorial about manifolds.
+
+## [0.15.23] 09/12/2024
+
+### Added
+
+* a field `point` to `ValidationFibreVector` to potentially store the point of the vector.
+* a field `store_base_point` to `ValidationManifold` to indicate whether for new fibre vectors the base point should be stored.
+* a keyword `ignore_contexts` to `ValidationManifold` to ignore certain contexts from validation, such as `:Input`, `:Output`, `:Point`, or `:Vector`.
+* a keyword `ignore_functions` to `ValidationFibreVector` to ignore certain contexts within a single function. This is provided as a dictionary with the key being the (allocating) function and the value is a context or vector of contexts.
+
+### Changed
+
+* the internal function `array_value` was renamed to `internal_value` and is now exported, since it can be also used on elements that store values different from arrays,
+e.g. a `ValidationMPoint` storing a subtype of a `ManifoldPoint`. `array_value` is hence deprecated.
+* Minimum Julia version is now 1.10 (the LTS which replaced 1.6)
 
 ## [0.15.22] 15/11/2024